Move common code to conftest
