+++ pynq_composable/test/test_mmio_partial_bitstreams.py
@@ -2,48 +2,11 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-from pynq import Overlay
 from pynq_composable.virtual import _mem_items
 import pytest
 
 
-@pytest.fixture
-def create_overlay():
-    for i in range(5):
-        try:
-            ol = Overlay("cv_dfx_3_pr.bit")
-        except OSError as e:
-            if "Bitstream file" in str(e):
-                pytest.exit(str(e))
-            elif i != 4:
-                continue
-            raise OSError("Could not program the FPGA")
-
-    yield ol
-    ol.free()
-
-
-@pytest.fixture
-def create_composable(create_overlay):
-    ol = create_overlay
-    cpipe = ol.composable
-    yield ol, cpipe
-
-
-def parameter():
-    ol = Overlay("cv_dfx_3_pr.bit")
-    cpipe = ol.composable
-    dfx_ip = list()
-    for k, v in cpipe.c_dict.unloaded.items():
-        element = {'ipname': k, 'modtype': v['modtype']}
-        dfx_ip.append(element)
-    return dfx_ip
-
-
-dfx_ip = parameter()
-
-
-@pytest.mark.parametrize('ip', dfx_ip)
+@pytest.mark.parametrize('ip', pytest.dfx_ip)
 def test_mmap_write(ip, create_composable):
     ol, cpipe = create_composable
     ipname, modtype = ip['ipname'], ip['modtype']
